requirements.txt
selenium==3.141.0
pytest==6.1.2
webdriver-manager==3.4.0
allure-pytest>=2.8.19
pytest-xdist==2.1.0
#monkeypatch==0.1rc3
pytest-json-report>=1.2.4
requests~=2.25.0
configparser~=5.0.1
PyVirtualDisplay==1.3.2
Appium-Python-Client==1.1.0
pytest-html-reporter==0.2.3
pytest-timeout==1.4.2
pytest-rerunfailures==10.1
httplib2==0.19.1
google-api-python-client==2.15.0
oauth2client==4.1.3
